fix: resolve tauri-specta BigIntForbidden and duplicate TS name errors
- src-tauri/src/lib.rs: export with BigIntExportBehavior::Number so
  i64/u64/usize fields map to TS number instead of bigint
- Prepend // @ts-nocheck to generated tauri.ts to suppress noUnusedLocals
  on the unused TAURI_CHANNEL/__makeEvents__ tail
- Broad usize → u32 across instrument-core I/O structs (api_key, crypto,
  csv, datetime, encoding, json, numbers, sql, text modules)
- regex-core: MatchResult / ExplainToken usize fields → u32
- json/diff + text/diff: add #[specta(rename = "JsonDiff…")] and
  #[specta(rename = "TextDiff…")] on AnnotatedLine / LineAnnotation to
  resolve duplicate TS names in generated tauri.ts
- Rebuild wasm-pkg to reflect instrument-core and regex-core changes

+++ src-core/instrument-core/src/crypto/md5.rs
@@ -27,7 +27,7 @@ pub struct Md5Output {
     /// 32-character hex hash.
     pub hash: String,
     /// Always 32 for MD5.
-    pub length: usize,
+    pub length: u32,
     /// Set when hashing fails (e.g. invalid UTF-8); normally None.
     pub error: Option<String>,
 }
